2d_barrel_matrix: Fix model double->float conv
The problem was that we used {__VA_ARGS__} previously that didn't allow
narrowing, but (__VA_ARGS__) was not working, since for empty list compiler was
getting confused if we are making some function declaration.

Solution is to use Model model __VA_ARGS__; but pass (lenght_scale) (surrounded
with parenthesis) as optional argument to macro.