CUDA opt: Skip kernel calc when voxel rho is 0
When measuring PSF where emission is from point sources, after several
iterations many previous rho voxels inside of event bounding box will be zero.

This optimization skips calculation for kernel (which is costly) in such case.
This is effective only when all voxels in warp are zero.

NOTE: This optimization will not likely improve performance in early iteration,
not for non-point sources.